+++ mopinion/settings.py
@@ -4,7 +4,7 @@ TOKEN_PATH = "/token"
 LATEST_VERSION = "2.0.0"
 
 # Some settings for dataclasses
-VERBOSITY_lEVELS = ["quiet", "normal", "full"]
-ITERATE_VERBOSITY_lEVELS = ["normal", "full"]
+VERBOSITY_LEVELS = ["quiet", "normal", "full"]
+ITERATE_VERBOSITY_LEVELS = ["normal", "full"]
 VERSIONS = ["1.18.14", "2.0.0"]
 CONTENT_NEGOTIATIONS = ["application/json", "application/x-yaml"]